blog_writer/requirements.txt
crewai>=0.28.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
langgraph>=0.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0
openai>=1.0.0 